app/requirements.txt
Django==4.1.3
django-filter
django-glrm==1.1.3
djangorestframework==3.14.0
django-tables2==2.4.1
Jinja2
psycopg2
yattag==1.14.0
celery
redis

